feat：能耗查询接口

+++ src/main/java/com/kfblue/seh/vo/DayTrendVO.java
@@ -4,6 +4,7 @@ import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Data
@@ -15,13 +16,13 @@ public class DayTrendVO {
     private List<DayValueVO> trends;
 
     @Schema(description = "当日用量")
-    private Double current;
+    private BigDecimal current;
 
     @Schema(description = "环比是与上一个相邻时间段的比较，环比增长率 =(当日 - 前一日) / 前一日 × 100%")
-    private Double momRate;
+    private BigDecimal momRate;
 
     @Schema(description = "同比是与去年同期的比较，同比增长率 =(今年当日 - 去年当日) / 去年当日 × 100%")
-    private Double yoyRate;
+    private BigDecimal yoyRate;
 
     @Schema(description = "单位")
     private String unit;
